requirements.txt
langchain==0.1.0
langchain-community==0.0.10
langchain-ollama==0.1.1
langchain-text-splitters==0.0.1
python-dotenv==1.0.0
pypdf==4.0.1
streamlit==1.28.1
faiss-cpu==1.7.4
ollama
EOF
